requirements.txt
anyio==3.6.1
bcrypt==4.0.1
click==8.1.3
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.85.0
greenlet==1.1.3.post0
h11==0.14.0
idna==3.4
passlib==1.7.4
psycopg2-binary==2.8.6
pydantic==1.10.2
PyJWT==2.5.0
python-dotenv==0.21.0
python-multipart==0.0.5
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.20.4
typing_extensions==4.4.0
uvicorn==0.18.3
